Rename implementations of DetectorControl to Controllers, restructure areadetector
