+++ src/ophyd_async/epics/areadetector/__init__.py
@@ -1,18 +1,27 @@
+from .controllers.pilatus_controller import PilatusController
+from .controllers.standard_controller import StandardController
 from .drivers.ad_driver import ADDriver
-from .nd_file_hdf import NDFileHDF
-from .nd_plugin import NDPlugin, NDPluginStats
+from .drivers.pilatus_driver import PilatusDriver
+from .pilatus import Pilatus
 from .single_trigger_det import SingleTriggerDet
 from .utils import FileWriteMode, ImageMode, ad_r, ad_rw
+from .writers.hdf_writer import HDFWriter
+from .writers.nd_file_hdf import NDFileHDF
+from .writers.nd_plugin import NDPlugin, NDPluginStats
 
 __all__ = [
+    "PilatusController",
+    "StandardController",
     "ADDriver",
-    "HDFStreamerDet",
-    "NDFileHDF",
-    "NDPlugin",
-    "NDPluginStats",
+    "PilatusDriver",
+    "Pilatus",
     "SingleTriggerDet",
     "FileWriteMode",
     "ImageMode",
     "ad_r",
     "ad_rw",
+    "HDFWriter",
+    "NDFileHDF",
+    "NDPlugin",
+    "NDPluginStats"
 ]
