+++ src/ophyd_async/epics/areadetector/controllers/pilatus_controller.py
@@ -1,14 +1,11 @@
 import asyncio
 
-from ophyd_async.core._signal.signal import (
-    set_and_wait_for_value,
-    wait_for_value,
-)
 from ophyd_async.core import AsyncStatus, DetectorControl, DetectorTrigger
 from ophyd_async.core._detector import DetectorControl, DetectorTrigger
+from ophyd_async.core._signal.signal import set_and_wait_for_value, wait_for_value
 from ophyd_async.epics.areadetector.utils import ImageMode
 
-from .drivers.pilatus_driver import PilatusDriver, TriggerMode
+from ..drivers.pilatus_driver import PilatusDriver, TriggerMode
 
 TRIGGER_MODE = {
     DetectorTrigger.internal: TriggerMode.internal,
@@ -17,7 +14,7 @@ TRIGGER_MODE = {
 }
 
 
-class PilatusControl(DetectorControl):
+class PilatusController(DetectorControl):
     def __init__(self, drv: PilatusDriver) -> None:
         self.driver = drv
