+++ src/ophyd_async/epics/areadetector/controllers/standard_controller.py
@@ -1,27 +1,26 @@
-import asyncio
 
-from ophyd_async.core import AsyncStatus, set_and_wait_for_value, wait_for_value
-from ophyd_async.core.detector.detector_control import DetectorControl
+from ophyd_async.core._detector.detector_control import DetectorControl
 from ophyd_async.core.utils import DEFAULT_TIMEOUT
 from ophyd_async.epics.areadetector.drivers.ad_driver import ADDriver, ImageMode
+from ophyd_async.core import AsyncStatus, set_and_wait_for_value, wait_for_value
+import asyncio
 
-
-class StandardControl(DetectorControl):
+class StandardController(DetectorControl):
     def __init__(self, drv: ADDriver) -> None:
         self.driver = drv
 
     async def get_deadtime(self, exposure: float) -> float:
         return 0.002
 
-    async def arm(self, num: int = 0) -> AsyncStatus:
+    async def arm(
+        self, num: int = 0
+    ) -> AsyncStatus:
         frame_timeout = DEFAULT_TIMEOUT + await self.driver.acquire_time.get_value()
         await asyncio.gather(
             self.driver.num_images.set(num),
             self.driver.image_mode.set(ImageMode.single),
         )
-        return await set_and_wait_for_value(
-            self.driver.acquire, True, timeout=frame_timeout
-        )
+        return await set_and_wait_for_value(self.driver.acquire, True, timeout=frame_timeout)
 
     async def disarm(self):
         # wait=False means don't caput callback. We can't use caput callback as we
