+++ src/ophyd_async/epics/areadetector/pilatus.py
@@ -1,11 +1,11 @@
-from ophyd_async.core import StaticDirectoryProvider, StandardDetector
+from ophyd_async.core import StandardDetector, StaticDirectoryProvider
 from ophyd_async.epics.areadetector.drivers.ad_driver import ADDriverShapeProvider
 from ophyd_async.epics.areadetector.writers.hdf_writer import HDFWriter as ADHDFWriter
 from ophyd_async.epics.areadetector.writers.nd_file_hdf import NDFileHDF
 from ophyd_async.epics.areadetector.writers.nd_plugin import NDPluginStats
 
+from .controllers.pilatus_controller import PilatusController
 from .drivers.pilatus_driver import PilatusDriver
-from .pilatus_control import PilatusControl
 
 dp = StaticDirectoryProvider("/dls/p45/data/cmxxx/i22-yyy", "i22-yyy-")
 
@@ -15,7 +15,7 @@ class Pilatus(StandardDetector):
         drv = PilatusDriver(prefix)
         hdf = NDFileHDF(prefix + "HDF:")
         super().__init__(
-            PilatusControl(drv),
+            PilatusController(drv),
             ADHDFWriter(
                 hdf, dp, lambda: self.name, ADDriverShapeProvider(drv), sum="NDStatsSum"
             ),
